In gibbs, convert tfrac to estimated read counts directly --- adds sampling variance

+++ src/SalmonQuantifyAlignments.cpp
@@ -1118,8 +1118,8 @@ bool processSample(AlignmentLibrary<ReadT>& alnLib,
         jointLog->info("Starting Gibbs Sampler");
         CollapsedGibbsSampler sampler;
         // The function we'll use as a callback to write samples
-        std::function<bool(const std::vector<int>&)> bsWriter =
-            [&gzw](const std::vector<int>& alphas) -> bool {
+        std::function<bool(const std::vector<double>&)> bsWriter =
+            [&gzw](const std::vector<double>& alphas) -> bool {
                 return gzw.writeBootstrap(alphas);
             };
